+++ src/Ormolu/Config.hs
@@ -1,7 +1,13 @@
+{-# LANGUAGE DeriveFunctor #-}
+{-# LANGUAGE RecordWildCards #-}
+
 -- | Configuration options used by the tool.
 module Ormolu.Config
   ( Config (..),
+    RegionIndices (..),
+    RegionDeltas (..),
     defaultConfig,
+    regionIndicesToDeltas,
     DynOption (..),
     dynOptionToLocatedStr,
   )
@@ -10,26 +16,66 @@ where
 import qualified SrcLoc as GHC
 
 -- | Ormolu configuration.
-data Config = Config
+data Config region = Config
   { -- | Dynamic options to pass to GHC parser
     cfgDynOptions :: ![DynOption],
     -- | Do formatting faster but without automatic detection of defects
     cfgUnsafe :: !Bool,
     -- | Output information useful for debugging
     cfgDebug :: !Bool,
-    -- | Checks if re-formatting the result is idempotent.
-    cfgCheckIdempotency :: !Bool
+    -- | Checks if re-formatting the result is idempotent
+    cfgCheckIdempotency :: !Bool,
+    -- | Region selection
+    cfgRegion :: !region
+  }
+  deriving (Eq, Show, Functor)
+
+-- | Region selection as the combination of start and end line numbers.
+data RegionIndices = RegionIndices
+  { -- | Start line of the region to format
+    regionStartLine :: !(Maybe Int),
+    -- | End line of the region to format
+    regionEndLine :: !(Maybe Int)
   }
   deriving (Eq, Show)
 
--- | Default 'Config'.
-defaultConfig :: Config
+-- | Region selection as the length of the literal prefix and the literal
+-- suffix.
+data RegionDeltas = RegionDeltas
+  { -- | Prefix length in number of lines
+    regionPrefixLength :: !Int,
+    -- | Suffix length in number of lines
+    regionSuffixLength :: !Int
+  }
+  deriving (Eq, Show)
+
+-- | Default @'Config' 'RegionIndices'@.
+defaultConfig :: Config RegionIndices
 defaultConfig =
   Config
     { cfgDynOptions = [],
       cfgUnsafe = False,
       cfgDebug = False,
-      cfgCheckIdempotency = False
+      cfgCheckIdempotency = False,
+      cfgRegion =
+        RegionIndices
+          { regionStartLine = Nothing,
+            regionEndLine = Nothing
+          }
+    }
+
+-- | Convert 'RegionIndices' into 'RegionDeltas'.
+regionIndicesToDeltas ::
+  -- | Total number of lines in the input
+  Int ->
+  -- | Region indices
+  RegionIndices ->
+  -- | Region deltas
+  RegionDeltas
+regionIndicesToDeltas total RegionIndices {..} =
+  RegionDeltas
+    { regionPrefixLength = maybe 0 (subtract 1) regionStartLine,
+      regionSuffixLength = maybe 0 (total -) regionEndLine
     }
 
 -- | A wrapper for dynamic options.